refactor: in 모듈에 있는 자바 객체 domain 모듈로 이동
