+++ core-domain/src/main/java/gohigher/user/User.java
@@ -1,5 +1,6 @@
 package gohigher.user;
 
+import gohigher.user.oauth2.Provider;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
