+++ out-adapter-persistence-jpa/src/main/java/gohigher/user/UserJpaEntity.java
@@ -1,5 +1,6 @@
 package gohigher.user;
 
+import gohigher.user.oauth2.Provider;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
 import jakarta.persistence.Enumerated;
